Updated to use long instead of int for volume field

+++ net.sourceforge.eclipsetrader.core/src/net/sourceforge/eclipsetrader/core/db/columns/Volume.java
@@ -55,7 +55,7 @@ public class Volume extends Column
         return 0;
     }
 
-    private int getValue(WatchlistItem item)
+    private long getValue(WatchlistItem item)
     {
         if (item.getSecurity() == null)
             return 0;
